pyproject.toml
[project]
name = "RLinf"
version = "0.2.0"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10,<=3.11.14"
keywords = [
    "reinforcement-learning",
    "embodied-intelligence",
    "large-language-models",
]
classifiers = [
    #   2 - Pre-Alpha
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: GPU :: NVIDIA CUDA :: 12 :: 12.4",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core System
    "accelerate",
    "ray[default]>=2.47.0",
    "torch==2.6.0",

    # Data processing
    "pylatexenc",
    "datasets==3.6.0",
    "latex2sympy2 @ git+https://github.com/RLinf/latex2sympy2.git",
    "sentencepiece",
    "torchdata",
    "wandb",
    "word2number",
    "regex",
    "scipy",

    # Utilities
    "hydra-core==1.4.0.dev1",
    "einops",
    "nvitop",
    "pybind11",
    "torch-memory-saver",
    "setuptools>=69.5.1,<75.9",
    "ninja",
    "pytest",
    "gsutil",
    "ruff",
    "pre-commit",
    "liger_kernel",
    "pip",
    "icmplib",
    "pyrealsense2",

    # Logging
    "swanlab>=0.6.11",
    "tensorboard",
]

[project.optional-dependencies]
sglang-vllm = [
    "transformers==4.51.1",
    "sglang[all]==0.4.6.post5",
    "vllm==0.8.5",
    # uvloop is not compatible with simulators, put it here
    "uvloop==0.21.0",
]
embodied = [
    "transformers==4.40.1",
    "draccus",
    "rich",
    "tensorflow_graphics",
    "peft==0.11.1",
    "timm==0.9.10",
    "sapien==3.0.1;platform_system=='Linux'",
    "tensordict",
    "imageio[ffmpeg]",
    "robosuite==1.4.1",
    "bddl",
    "easydict",
    "cloudpickle",
    "gym",
    "av",
]
franka = [
    "pyyaml",
    "rospkg",
    "empy==3.3.4",
    "gymnasium==0.29.1",
    "pyrealsense2",
    "pyquaternion",
    "opencv-python",
    "defusedxml",
    "pyspacemouse", 
    "imageio[ffmpeg]",
]

[tool.uv]
prerelease = "allow"
conflicts = [
    [
      { extra = "sglang-vllm" },
      { extra = "embodied" },
      { extra = "franka" }
    ],
]
override-dependencies = [
    "torch==2.6.0",
    "torchvision==0.21.0",
    "torchaudio==2.6.0",
    "torchcodec==0.2",
    "xgrammar==0.1.19",
    "uvloop==0.21.0",
    "tensorflow-addons==0.23.0",
    "hydra-core==1.4.0.dev1",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
isort = {known-first-party = ["rlinf"]}
select = ["C", "E", "F", "I", "W", "CPY001", "RUF013", "UP006", "PERF102", "PLC1802", "PLC0208", "D", "RUF002", "DOC102", "DOC202", "DOC403", "DOC502"]
ignore = [
    "C901", # "complex-structure"
    "E501", # "line-too-long"
    "E741", # "ambiguous-variable-name"
    "D100", # "Missing docstring in public module"
    "D104", # "Missing docstring in public package"
    "D203", # "incorrect-blank-line-before-class", conflict with D211,
    "D213", # "multi-line-summary-second-line", conflict with D212
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
# Only enable docstring check for the scheduler module for now.
"!rlinf/scheduler/**.py" = ["D", "RUF002"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"